+++ docs/source/changelog.rst
@@ -17,8 +17,10 @@ Array
 DataFrame
 +++++++++
 
--  Support month timedeltas in repartition(freq=...) (:pr:`3110`) `Matthew Rocklin`_
--  Avoid mutation in dataframe groupby tests (:pr:`3118`) `Matthew Rocklin`_
+- Support month timedeltas in repartition(freq=...) (:pr:`3110`) `Matthew Rocklin`_
+- Avoid mutation in dataframe groupby tests (:pr:`3118`) `Matthew Rocklin`_
+- ``read_csv``, ``read_table``, and ``read_parquet`` accept iterables of paths
+  (:pr:`3124`) `Jim Crist`_
 
 Bag
 +++